requirements.txt
alembic==1.12.0
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
gunicorn==22.0.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.4.0
python-engineio==4.7.1
python-socketio==5.9.0
requests==2.31.0
simple-websocket==0.10.1
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.0.6
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.0.1
proglog==0.1.10
psycopg2-binary
